copy text refactor
